+++ packages/cli/__tests__/options/version.test.ts
@@ -1,14 +1,9 @@
 import { describe, expect, it } from "vitest";
-import { execFileSync } from "node:child_process";
-import { fileURLToPath } from "node:url";
 import packageJson from "../../package.json" with { type: "json" };
+import { getCliVersion } from "../../src/options/version.js";
 
-describe("ao --version", () => {
+describe("getCliVersion", () => {
   it("matches the CLI package version", () => {
-    const tsxEntry = fileURLToPath(new URL("../../node_modules/.bin/tsx", import.meta.url));
-    const cliEntry = fileURLToPath(new URL("../../src/index.ts", import.meta.url));
-    const output = execFileSync(tsxEntry, [cliEntry, "--version"], { encoding: "utf8" }).trim();
-
-    expect(output).toBe(packageJson.version);
+    expect(getCliVersion()).toBe(packageJson.version);
   });
 });
