+++ packages/cli/__tests__/program.test.ts
@@ -1,0 +1,9 @@
+import { describe, expect, it } from "vitest";
+import packageJson from "../package.json" with { type: "json" };
+import { createProgram } from "../src/program.js";
+
+describe("createProgram", () => {
+  it("uses the CLI package version", () => {
+    expect(createProgram().version()).toBe(packageJson.version);
+  });
+});
