+++ packages/cli/src/index.ts
@@ -1,56 +1,5 @@
 #!/usr/bin/env node
 
-import { Command } from "commander";
-import { createRequire } from "node:module";
-import { registerInit } from "./commands/init.js";
-import { registerStatus } from "./commands/status.js";
-import { registerSpawn, registerBatchSpawn } from "./commands/spawn.js";
-import { registerSession } from "./commands/session.js";
-import { registerSend } from "./commands/send.js";
-import { registerReviewCheck } from "./commands/review-check.js";
-import { registerDashboard } from "./commands/dashboard.js";
-import { registerOpen } from "./commands/open.js";
-import { registerStart, registerStop } from "./commands/start.js";
-import { registerLifecycleWorker } from "./commands/lifecycle-worker.js";
-import { registerVerify } from "./commands/verify.js";
-import { registerDoctor } from "./commands/doctor.js";
-import { registerUpdate } from "./commands/update.js";
-import { registerSetup } from "./commands/setup.js";
-import { registerPlugin } from "./commands/plugin.js";
-import { getConfigInstruction } from "./lib/config-instruction.js";
+import { createProgram } from "./program.js";
 
-const require = createRequire(import.meta.url);
-const packageJson = require("../package.json") as { version: string };
-const program = new Command();
-
-program
-  .name("ao")
-  .description("Agent Orchestrator — manage parallel AI coding agents")
-  .version(packageJson.version);
-
-registerInit(program);
-registerStart(program);
-registerStop(program);
-registerStatus(program);
-registerSpawn(program);
-registerBatchSpawn(program);
-registerSession(program);
-registerSend(program);
-registerReviewCheck(program);
-registerDashboard(program);
-registerOpen(program);
-registerLifecycleWorker(program);
-registerVerify(program);
-registerDoctor(program);
-registerUpdate(program);
-registerSetup(program);
-registerPlugin(program);
-
-program
-  .command("config-help")
-  .description("Show config schema and guide for creating agent-orchestrator.yaml")
-  .action(() => {
-    console.log(getConfigInstruction());
-  });
-
-program.parse();
+createProgram().parse();
